requirements.txt
environs==9.5.0
Django==4.1.7
django-extensions==3.2.1
phonenumbers==8.13.8
django-phonenumber-field==7.0.2
nameparser==1.1.2
pymorphy2==0.9.1
djangorestframework==3.14.0
Pillow==9.5.0
more-itertools==9.1.0
django-debug-toolbar==4.0.0
gunicorn==20.1.0
psycopg2-binary==2.9.6
simplejson==3.19.1
logtail-python==0.2.6
num2words==0.5.12
lxml==4.9.3
xhtml2pdf==0.2.11
xlwt==1.3.0
openpyxl==3.1.2
django-filter==23.2
redis==5.0.0
schedule==1.1.0
django-meta==2.4.0
